% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{calc_adj_frag}
\alias{calc_adj_frag}
\title{Calculate adjacency matrices and fragility heatmap from iEEG recording}
\usage{
calc_adj_frag(ieegts, t_window, t_step, lambda = NULL)
}
\arguments{
\item{ieegts}{Numeric. A matrix of iEEG time series,
with time points as rows and electrodes names as columns}

\item{t_window}{Integer. The number of time points to use in each window}

\item{t_step}{Integer.The number of time points to move the window each time}

\item{lambda}{Numeric. The lambda value to use in the ridge regression.
If NULL, the lambda will be chosen automatically
ensuring that AI is table}
}
\value{
A list containing the normalized ieegts,
adjacency matrices, fragility, and R^2 values
}
\description{
For each time window, a discrete stable Linear time system
(adjacency matrix) is computed named Ai
Lambda=NULL ensures that the matrix is stable
}
\examples{
data <- matrix(rnorm(100), nrow = 10)
t_window <- 10
t_step <- 5
lambda <- 0.1
calc_adj_frag(ieegts = data, t_window = t_window, t_step = t_step, lambda = lambda)

data("ptEpoch")
t_window <- 250
t_step <- 125
lambda <- NULL
ptEpochSmall<-ptEpoch[9900:12001,]
resfragcalc_adj_frag(ieegts = ptEpochSmall, t_window = t_window, t_step = t_step, lambda = lambda)

}
