% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FragStat.r, R/class_Fragility.r
\name{$,FragStat-method}
\alias{$,FragStat-method}
\alias{$<-,FragStat-method}
\alias{$,Fragility-method}
\alias{$<-,Fragility-method}
\title{Getters and Setters for S4 object}
\usage{
\S4method{$}{FragStat}(x, name)

\S4method{$}{FragStat}(x, name) <- value

\S4method{$}{Fragility}(x, name)

\S4method{$}{Fragility}(x, name) <- value
}
\arguments{
\item{x}{S4 object}

\item{name}{Slot name}

\item{value}{Value to set}
}
\value{
S4 object itself or slot value
}
\description{
Getters and Setters for S4 object
}
