% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{checkIndex}
\alias{checkIndex}
\title{Check and filter display index}
\usage{
checkIndex(display, elecNames)
}
\arguments{
\item{display}{Numeric or character. Display index}

\item{elecNames}{Character. All electrode names}
}
\description{
Check and filter display index
}
