% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Fragility.r
\name{nrow,Fragility-method}
\alias{nrow,Fragility-method}
\alias{ncol,Fragility-method}
\title{Get the number of rows or columns of a Fragility object}
\usage{
\S4method{nrow}{Fragility}(x)

\S4method{ncol}{Fragility}(x)
}
\arguments{
\item{x}{A Fragility object}
}
\value{
\itemize{
\item \code{nrow(x)}: The number of rows (electrodes) in the fragility matrix.
\item \code{ncol(x)}: The number of columns (time points) in the fragility matrix.
\item \code{dim(x)}: A vector of length 2 containing the number of rows and columns in the fragility matrix.
}
}
\description{
Get the number of rows or columns of a Fragility object
}
