% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Epoch.r
\docType{class}
\name{Epoch-class}
\alias{Epoch-class}
\alias{.Epoch}
\title{Epoch Class}
\description{
S4 class to handle epoch data with electrodes and time points
}
\section{Slots}{

\describe{
\item{\code{data}}{a tibble containing epoch data (columns=time points, rows=electrodes)}

\item{\code{times}}{Numeric vector containing time range}
}}

