% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Epoch.r
\name{truncateTime}
\alias{truncateTime}
\alias{truncateTime,Epoch-method}
\alias{$,Epoch-method}
\alias{$<-,Epoch-method}
\alias{[,Epoch-method}
\alias{nrow,Epoch-method}
\alias{ncol,Epoch-method}
\alias{colnames,Epoch-method}
\alias{colnames<-,Epoch-method}
\alias{rownames,Epoch-method}
\alias{rownames<-,Epoch-method}
\alias{names,Epoch-method}
\alias{names<-,Epoch-method}
\alias{as.matrix,Epoch-method}
\alias{as.data.frame,Epoch-method}
\title{Epoch Methods}
\usage{
truncateTime(x, from, to)

\S4method{truncateTime}{Epoch}(x, from, to)

\S4method{$}{Epoch}(x, name)

\S4method{$}{Epoch}(x, name) <- value

\S4method{[}{Epoch}(x, i, j)

\S4method{nrow}{Epoch}(x)

\S4method{ncol}{Epoch}(x)

\S4method{colnames}{Epoch}(x)

\S4method{colnames}{Epoch}(x) <- value

\S4method{rownames}{Epoch}(x)

\S4method{rownames}{Epoch}(x) <- value

\S4method{names}{Epoch}(x)

\S4method{names}{Epoch}(x) <- value

\S4method{as.matrix}{Epoch}(x)

\S4method{as.data.frame}{Epoch}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Epoch object}

\item{from}{Numeric value specifying start of new time range}

\item{to}{Numeric value specifying end of new time range}

\item{name}{a value name, must be one of 'electrodes', 'times', 'timeRange', 'matrix'}

\item{value}{Value to set}

\item{i}{Row (time) indices}

\item{j}{Column (electrode) indices}

\item{...}{additional arguments}
}
\value{
truncateTime: Truncated object

nrow: Number of rows in the data

ncol: Number of columns in the data

colnames: electrode names of the data

rownames: time points of the data

names: electrode names of the data
}
\description{
\code{truncateTime}: Truncating time range

\code{$}: get Epoch object properties, must be one of 'electrodes',
'times', 'timeRange', 'matrix'

\verb{$<-}: set Epoch properties, must be one of 'electrodes',
'times', 'timeRange', 'matrix'

\code{[}: Subset an Epoch object using matrix indexing syntax

\code{nrow}, \code{ncol}, \code{colnames}, \code{rownames}, \code{names}: Getting the data properties,
similar to base R functions.
}
