% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Epoch.r
\name{Epoch}
\alias{Epoch}
\title{Constructor for Epoch class}
\usage{
Epoch(data, electrodes = NULL, timeRanges = NULL, times = NULL)
}
\arguments{
\item{data}{Matrix containing epoch data (rows=electrodes, columns=time points)}

\item{electrodes}{Optional character vector for electrode names, if not provided, column names of data are used. If both are NULL, electrodes are named E1, E2, ...}

\item{timeRanges}{Optional numeric vector of 2 containing start
and end time points. Only one of times or timeRanges can be non-null}

\item{times}{Optional numeric vector of time points. Only one of times or
timeRanges can be non-null}
}
\value{
An Epoch object
}
\description{
Constructor for Epoch class
}
