% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{estimateSOZ}
\alias{estimateSOZ}
\title{Find Seizure Onset Zone}
\usage{
estimateSOZ(
  x,
  method = c("mean", "median", "max", "min"),
  proportion = 0.1,
  ...
)
}
\arguments{
\item{x}{Fragility object}

\item{method}{Character. The method to use to find the onset zone.
Must be one of 'max', 'min', or "mean"}

\item{proportion}{Numeric. The proportion of electrodes to consider as the onset zone.
The electrode number will be rounded to the nearest integer.}

\item{...}{Additional arguments}
}
\value{
A vector of electrode names, or indices if the electrode names are NULL
}
\description{
The function estimates the seizure onset zone (SOZ). For each row, it calculates the maximum, minimum, or mean of row. The rows with the highest values are considered as the SOZ.
}
