% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{findOnset}
\alias{findOnset}
\title{Find Serzure Onset Zone}
\usage{
findOnset(x, method = c("mean", "max", "min"), proportion = 0.1, ...)
}
\arguments{
\item{x}{Fragility object}

\item{method}{Character. The method to use to find the onset zone.
Must be one of 'max', 'min', or "mean"}

\item{proportion}{Numeric. The proportion of electrodes to consider as the onset zone.
The electrode number will be rounded to the nearest integer.}

\item{...}{Additional arguments}
}
\value{
A vector of electrode names, or indices if the electrode names are NULL
}
\description{
Find Serzure Onset Zone
}
