% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{FourthOrderButterworthFilter}
\alias{FourthOrderButterworthFilter}
\title{Band-pass filter voltage between 0.5 and Nyquist frequency with fourth-order Butterworth filter}
\usage{
FourthOrderButterworthFilter(epoch)
}
\arguments{
\item{epoch}{}
}
\description{
Band-pass filter voltage between 0.5 and Nyquist frequency with fourth-order Butterworth filter
}
\examples{
epochfilt<-FourthOrderButterworthFilter(epoch)
}
