% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility.r
\name{frag_quantile}
\alias{frag_quantile}
\title{Compute quantiles, mean and standard deviation for two electrodes group marked as soz non marked as soz}
\usage{
frag_quantile(ieegts, frag, t_step, ElectrodesData, fs, time_window_ictal)
}
\arguments{
\item{frag}{fragility matrix}

\item{t_step}{fragility step size}

\item{ElectrodesData}{Electrodes data}

\item{fs}{Acquisition frequency}

\item{time_window_ictal}{ictal time window}

\item{ieegs}{ictal signal}
}
\value{
quantile matrix, mean and sdv from both electrodes groups
}
\description{
Compute quantiles, mean and standard deviation for two electrodes group marked as soz non marked as soz
}
\examples{
data("fragm3sp5s")
data("pt01Epochm3sp5s")
data("ElectrodesDataPT01")
ieegts=pt01Epochm3sp5s
frag=fragm3sp5s
fs=1000
time_window_ictal=c(-10,10)
t_step=150
ElectrodesData=ElectrodesDataPT01

}
