% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragStat.r
\name{fragilityRow}
\alias{fragilityRow}
\title{Compute the normalized fragility row for adjacency matrix A}
\usage{
fragilityRow(A, nSearch = 100, normalize = TRUE)
}
\arguments{
\item{A}{Numeric. Adjacency Matrix}

\item{nSearch}{Integer. Number of eigenvalues tried to find the minimum norm vector}

\item{normalize}{Logical. If TRUE, the fragility row is normalized}
}
\description{
The matrix A is used for the regression: A * x(t) = x(t+1)
}
