% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IM_main.R
\name{GetAdjFrag}
\alias{GetAdjFrag}
\title{Replaces calc_adj_frag (I don't like snake_case - I prefer CamelCase)}
\usage{
GetAdjFrag(ieegts, t_window, t_step, lambda = NULL, nSearch = 10L)
}
\arguments{
\item{ieegts}{Numeric. A matrix of iEEG time series x(t),
with time points as rows and electrodes names as columns}

\item{t_window}{Integer. The number of time points to use in each window}

\item{t_step}{Integer. The number of time points to move the window each time}

\item{lambda}{Numeric. The lambda value to use in the ridge regression.}

\item{nSearch}{Integer. Number of minimization to compute the fragility row}
}
\value{
Fragility obj with adjacency matrices, fragility, and R^2 values
}
\description{
Calculate adjacency matrices and fragility matrix from iEEG recording
}
\examples{
# A simple example
GetAdjFrag(matrix(rnorm(100), 10), 10, 5, .1)
# A more realistic example, but it will take a while to run
\dontrun{
resfrag <- GetAdjFrag(pt01Epoch[9001:12000,], 250, 125, NULL, 100)
}
}
