% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmap_frag}
\alias{heatmap_frag}
\title{Visualization}
\usage{
heatmap_frag(
  frag,
  goodelec,
  sozelec,
  ieegts,
  option = NULL,
  plotelec,
  subject_code,
  j
)
}
\arguments{
\item{frag}{fragility matrix results}

\item{goodelec}{}

\item{sozelec}{}

\item{ieegts}{Numeric. A matrix of iEEG time series x(t),
with time points as rows and electrodes names as columns}
}
\description{
plot fragility heatmaps with electrodes marked as soz colored
}
\examples{
heatmap_frag(frag=frag,goodelec=goodelec,sozelec=sozelec,ieegts=ptEpochm1sp2s,subject_code='pt01',j=1)
}
