% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmapFrag}
\alias{heatmapFrag}
\title{Visualization functions (raw signal, fragility matrix)}
\usage{
heatmapFrag(frag, sozIndex = NULL)
}
\arguments{
\item{frag}{Fragility object from \code{calcAdjFrag}}

\item{sozIndex}{Integer or string. A group of electrodes to mark as in the Seizure Onset Zone (SOZ)}

\item{title}{String. Figure title}

\item{display}{Integer or string. Vector electrodes to display}
}
\value{
Heatmap plot of the fragility matrix with soz electrodes in blue in the bottom
}
\description{
plot fragility heatmaps with electrodes marked as soz colored
}
\examples{
# use integer index for display and soz electrodes
data("pt01Epoch")
sozIndex <- attr(pt01Epoch, "sozIndex")
data("pt01Frag")
timeRange <- c(-1, 2)
display <- c(sozIndex, 77:80)
fragplot <- heatmapFrag(
    frag = pt01Frag, sozID = sozIndex,
    timeRange = timeRange, title = "PT01 seizure 1", display = display
)
fragplot


# use electrodes name for display and soz electrodes
data("pt01Epoch")
sozNames <- attr(pt01Epoch, "sozNames")
data("pt01Frag")
timeRange <- c(-1, 2)
display <- c(sozNames, "MLT1", "MLT2", "MLT3", "MLT4")
fragplot <- heatmapFrag(
    frag = pt01Frag, sozID = sozNames,
    timeRange = timeRange, title = "PT01 seizure 1", display = display
)
fragplot

# save plot to file with ggplot2
data("pt01Epoch")
data("pt01Frag")
sozIndex <- attr(pt01Epoch, "sozIndex")
timeRange <- c(-10, 10)
display <- c(sozIndex, 77:80)
pathplot <- "~"
title <- "PT01sz1"
resfile <- paste(pathplot, "/FragilityHeatMap", title, ".png", sep = "")
fragplot <- heatmapFrag(
    frag = pt01Frag, sozID = sozIndex, timeRange = timeRange,
    title = title, display = display
)
fragplot
ggplot2::ggsave(resfile)

}
