% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotFragDistribution}
\alias{plotFragDistribution}
\title{Plot Fragility time distribution for two electrodes group marked and non-marked as soz}
\usage{
plotFragDistribution(frag, sozIndex = NULL)
}
\arguments{
\item{frag}{Fragility object from \code{calcAdjFrag}}

\item{sozIndex}{Integer or string. A group of electrodes to mark as in the Seizure Onset Zone (SOZ)}

\item{stat}{FragStat object, a FragStat object from \code{fragStat}}

\item{title}{String. Figure title}
}
\value{
plot fragility distribution
}
\description{
Plot Fragility time distribution for two electrodes group marked and non-marked as soz
}
\examples{
data("pt01Epoch")
sozindex <- attr(pt01Epoch, "sozindex")
# Load the precomputed fragility object
timeRange <- c(-10, 10)
data("pt01Frag")
# compute fragility statistics evolution with time (mean and standard deviation) for soz and
# non soz groups
pt01fragstat <- fragStat(frag = pt01Frag, sozID = sozindex)
# plot the statistical results
pfragstat <- plotFragDistribution(stat = pt01fragstat, timeRange = timeRange)
pfragstat
}
