% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotFragHeatmap}
\alias{plotFragHeatmap}
\alias{plotFragQuantile}
\alias{plotFragDistribution}
\title{Visualization functions (raw signal, fragility matrix)}
\usage{
plotFragHeatmap(frag, groupIndex = NULL)

plotFragQuantile(frag, groupIndex = NULL, groupName = "SOZ")

plotFragDistribution(frag, groupIndex = NULL, groupName = "SOZ")
}
\arguments{
\item{frag}{Fragility object from \code{calcAdjFrag}}

\item{groupIndex}{Integer or string. A group of electrodes to mark}
}
\value{
A ggplot object
}
\description{
\code{plotFragHeatmap}: plot fragility heatmaps with electrodes marked as soz colored

\code{plotFragQuantile}: Plot Fragility time quantiles for two electrodes group marked as SOZ and reference

\code{plotFragQuantile}: Plot Fragility time distribution for two electrodes group marked as SOZ and reference
}
\examples{

data("pt01EcoG")

## sozNames is the name of the electrodes we assume are in the SOZ
sozNames <- metaData(pt01EcoG)$sozNames

## precomputed fragility object
data("pt01Frag")

## plot the fragility heatmap
plotFragHeatmap(frag = pt01Frag, groupIndex = sozNames)

## plot the fragility quantiles
plotFragQuantile(frag = pt01Frag, groupIndex = sozNames)

## plot the fragility distribution
plotFragDistribution(frag = pt01Frag, groupIndex = sozNames)

}
