% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pt01EcoG}
\alias{pt01EcoG}
\alias{pt01Frag}
\title{Pt01 seizure 1 around seizure onset}
\format{
pt01EcoG: A Matrix with 84 rows (electrodes) and 3000 columns (time points)

pt01Frag: A fragility object results of applying the main function \code{calcAdjFrag} to pt01EcoG with \code{window} = 250 and \code{step} = 125
}
\source{
Fragility Multi-Center Retrospective Study
(\href{https://openneuro.org/datasets/ds003029/versions/1.0.0}{OpenNeuro})
}
\usage{
## EEG data
data(pt01EcoG)
}
\description{
This data corresponds to the first seizure of patient from
the Fragility Data Set. EcoG recording gathered in collaboration with the National Institute of Health. The data contains only the good channels.
It has been notch filtered and common average referenced in RAVE.
The time range for full data is (-10:10s). Due to the size limit of the package, The full data has been epoched -1:2s around the seizure onset.
The acquisition frequency is 1000 Hz
}
\keyword{datasets}
