% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pt01Epochm1sp2s}
\alias{pt01Epochm1sp2s}
\alias{pt01Fragm1sp2s}
\title{Pt01 seizure 1 (-10:10s) around seizure onset}
\format{
pt01Epochm1sp2s: A Matrix with 3000 rows (time points) and 84 columns (electrodes)

pt01Fragm1sp2s: A fragility object results of applying the main function \code{calcAdjFrag}
to pt01Epochm1sp2s
}
\source{
Fragility Multi-Center Retrospective Study
(\href{https://openneuro.org/datasets/ds003029/versions/1.0.0}{OpenNeuro})
}
\usage{
## EEG data
data(pt01Epochm1sp2s)
}
\description{
This data corresponds to the first seizure of patient PT01 from
the Fragility Data Set.
The data contains only the good channels.
It has been notch filtered and common average referenced in RAVE.
The full data has been epoched -1:2s around the seizure onset.
The acquisition frequency is 1000 Hz
EcoG recording gathered in collaboration with the National Institue of Health
}
\keyword{datasets}
