% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigde.r
\name{ridge}
\alias{ridge}
\title{Ridge Regression for Electrode Readings}
\usage{
ridge(xt, xtp1, lambda, intercept = FALSE, iw)
}
\arguments{
\item{xt}{matrix. iEEG time series for a given window}

\item{xtp1}{matrix. the iEEG time serie at the next time point}

\item{lambda}{}
}
\description{
Ridge regression to compute matrix A such as A xt = xtpt1
}
\examples{
...

}
