% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigde.r
\name{ridgeR2}
\alias{ridgeR2}
\title{computes R2}
\usage{
ridgeR2(xt, xtp1, A)
}
\arguments{
\item{xt}{matrix. iEEG time series for a given window,
with electrodes names as rows and time points as columns}

\item{xtp1}{matrix. the iEEG time serie at the next time point,
with electrodes names as rows and time points as columns}

\item{A}{adjacency matrix}
}
\description{
computes R2
}
