% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigde.r
\name{ridgeSearch}
\alias{ridgeSearch}
\title{Ridge Regression for Electrode Readings}
\usage{
ridgeSearch(xt, xtp1, lambda = NULL)
}
\arguments{
\item{xt}{matrix. iEEG time series for a given window,
with electrodes names as rows and time points as columns}

\item{xtp1}{matrix. the iEEG time serie at the next time point,
with electrodes names as rows and time points as columns}

\item{lambda}{Numeric Vector. A user supplied lambda sequence.}
}
\value{
adjacency matrix Afin with lambda as attribute
}
\description{
Ridge regression to compute matrix adjancency matrix A such as A xt = xtpt1
the lambda parmeter is found by dichotomy such that A is stable
(all eigenvalues have a norm less than one)
}
