% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigde.r
\name{ridgesearchlambdadichomotomy}
\alias{ridgesearchlambdadichomotomy}
\title{Title}
\usage{
ridgesearchlambdadichomotomy(xt, xtp1, intercept = FALSE, iw)
}
\arguments{
\item{xt}{}

\item{xtp1}{}

\item{intercept}{}

\item{iw}{}
}
\description{
Title
}
