% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Fragility.r
\name{show,Fragility-method}
\alias{show,Fragility-method}
\title{Print the Fragility object}
\usage{
\S4method{show}{Fragility}(object)
}
\arguments{
\item{object}{A Fragility object}
}
\value{
the object itself
}
\description{
Print the Fragility object
}
