% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FragStat.r
\name{show,FragStat-method}
\alias{show,FragStat-method}
\title{Print the FragStat object}
\usage{
\S4method{show}{FragStat}(object)
}
\arguments{
\item{object}{A FragStat object}
}
\value{
the object itself
}
\description{
Print the FragStat object
}
