% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visu_iEEG_data}
\alias{visu_iEEG_data}
\title{Visualization of ictal iEEG}
\usage{
visu_iEEG_data(
  ieegts,
  time_window = NULL,
  title = "Patient name seizure number",
  display = NULL
)
}
\arguments{
\item{ieegts}{Matrix or Fragility object. Either a matrix of iEEG time
series x(t), with time points as rows and electrodes names as columns,
or a Fragility object from \code{calc_adj_frag}}

\item{time_window}{Numeric Vector of length 2. The time window to display at the x-axis}

\item{title}{String. Figure title}

\item{display}{Integer or string. Vector electrodes to display}
}
\value{
plot raw signal
}
\description{
Visualization of ictal iEEG
}
\examples{
data("pt01Epoch")
sozindex <- attr(pt01Epoch,"sozindex")
## We will use data from -3s to 5s around the seizure onset
pt01Epochm3sp5s<-pt01Epoch[7001:15000,]
display <- c(sozindex,77:80)
time_window <- c(-3,5)
iEEGplot<-visu_iEEG_data(ieegts=pt01Epochm3sp5s,time_window=time_window,display=display)
iEEGplot
}
