% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visuIEEGData}
\alias{visuIEEGData}
\title{Visualization of ictal iEEG}
\usage{
visuIEEGData(epoch, groupIndex = NULL, maxLabels = 50)
}
\arguments{
\item{epoch}{Matrix or Epoch object. iEEG data matrix or Epoch object. If matrix, the row names are the electrode names and the column names are the time points}

\item{groupIndex}{Integer or string. A group of electrodes to mark}

\item{maxLabels}{Integer. Maximum number of labels to show on y-axis. Default is 50. The actual number of labels may be less than this value if there are too many electrodes.}
}
\value{
A ggplot object
}
\description{
Visualization of ictal iEEG
}
\examples{
data("pt01EcoG")

## Visualize a subject of electrodes
sozIndex <- which(rowData(pt01EcoG)$soz)
display <- c(sozIndex, 77:80)

visuIEEGData(epoch = pt01EcoG[display, ])
}
