% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{checkIndex}
\alias{checkIndex}
\title{Check and keep valid index only}
\usage{
checkIndex(indices, names)
}
\arguments{
\item{indices}{Numeric or character index to check}

\item{names}{Character. All names corresponding to the indices}
}
\description{
Check and keep valid index only
}
