% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Fragility.r
\name{nrow,Fragility-method}
\alias{nrow,Fragility-method}
\title{Get the number of rows or columns of a Fragility object}
\usage{
\S4method{nrow}{Fragility}(x)
}
\arguments{
\item{x}{A Fragility object}
}
\description{
Get the number of rows or columns of a Fragility object
}
