% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility.r
\name{frag_quantile}
\alias{frag_quantile}
\title{Compute  quantile for two electrodes group marked as soz non marked as soz}
\usage{
frag_quantile(frag, elecsoz = NULL)
}
\arguments{
\item{frag}{Numeric. Fragility matrix. Row Electrode names. Column index}

\item{elecsoz}{Integer.  Vector soz electrodes (for good electrodes)}
}
\value{
quantile matrix for the two groups. Row is the quantile. Column is time
}
\description{
Compute  quantile for two electrodes group marked as soz non marked as soz
}
\examples{
data("fragm3sp5s")
sozindex<-attr(fragm3sp5s,"sozindex")
# compute fragility statistics evolution with time (mean and standard deviation) for soz and
# non soz groups
qmatsozsozc=frag_quantile(frag=fragm3sp5s, elecsoz=sozindex)
}
