% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility.r
\name{frag_stat}
\alias{frag_stat}
\title{Compute  mean and standard deviation for two electrodes group marked as soz non marked as soz}
\usage{
frag_stat(frag, elecsoz = NULL)
}
\arguments{
\item{frag}{Numeric. Fragility matrix. Row Electrode names. Column index}

\item{elecsoz}{Integer.  Vector soz electrodes (for good electrodes)}
}
\value{
list of 4 items with mean and sdv from both electrodes groups
}
\description{
Compute  mean and standard deviation for two electrodes group marked as soz non marked as soz
}
\examples{
data("fragm3sp5s")
sozindex<-attr(fragm3sp5s,"sozindex")
fragstat=frag_stat(frag=fragm3sp5s, elecsoz=sozindex)
}
