% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility.r
\name{frag_stat}
\alias{frag_stat}
\title{Compute quantiles, mean and standard deviation for two electrodes group marked as soz non marked as soz}
\usage{
frag_stat(frag, elecsoz)
}
\arguments{
\item{frag}{Matrix or Fragility object. Either a matrix with row as Electrode names and Column as fragility index, or a Fragility object from \code{calc_adj_frag}}

\item{elecsoz}{Integer.  Vector soz electrodes (for good electrodes)}
}
\value{
list of 5 items with quantile matrix, mean and sdv from both electrodes groups
}
\description{
Compute quantiles, mean and standard deviation for two electrodes group marked as soz non marked as soz
}
\examples{
data("fragm3sp5s")
data("elecsoz")
fragstat<-frag_stat(frag=fragm3sp5s, elecsoz=elecsoz)
}
