% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility.r
\name{fragilityRow}
\alias{fragilityRow}
\title{Compute the fragility row for adjacency matrix A}
\usage{
fragilityRow(A, nSearch = 100)
}
\arguments{
\item{A}{Numeric. Adjacency Matrix}

\item{nSearch}{Integer. Number of eigenvalues tried to find the minimum norm vector}
}
\description{
Compute the fragility row for adjacency matrix A
}
