% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility.r
\name{fragilityRowNormalized}
\alias{fragilityRowNormalized}
\title{Compute the normalized fragility row for adjacency matrix A}
\usage{
fragilityRowNormalized(A, nSearch = 100)
}
\arguments{
\item{A.}{Numeric. Adjacency Matrix}

\item{nSearch.}{Integer. Number of eigenvalues tried to find the minimum norm vector}
}
\description{
Compute the normalized fragility row for adjacency matrix A
}
