% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility.r
\name{fragilityRowNormalized}
\alias{fragilityRowNormalized}
\title{Compute the normalized fragility row for adjacency matrix A}
\usage{
fragilityRowNormalized(A, nSearch = 100)
}
\arguments{
\item{A}{}

\item{nSearch}{}
}
\description{
Compute the normalized fragility row for adjacency matrix A
}
