% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmap_frag}
\alias{heatmap_frag}
\title{Visualization functions (raw signal, fragility matrix)}
\usage{
heatmap_frag(
  frag,
  elecsoz,
  time_window,
  title = "PT01 seizure 1",
  display = NULL,
  pathplot = ""
)
}
\arguments{
\item{frag}{Numeric. Fragility matrix results}

\item{elecsoz}{Integer. Vector soz electrodes (for good electrodes)}

\item{time_window}{Numeric Vector. Fragility heatmap time window around seizure onset (s)}

\item{title}{String. Figure title}

\item{display}{Integer. Electrodes to display}

\item{pathplot}{String. File Path to save plot}
}
\value{
Heatmap plot of the fragility matrix with soz electrodes in blue in the bottom
}
\description{
plot fragility heatmaps with electrodes marked as soz colored
}
\examples{
data("fragm3sp5s")
data("elecsoz")
time_window=c(-3:5)
display=c(elecsoz,77:80)
heatmap_frag(frag=fragm3sp5s,elecsoz=elecsoz,time_window=c(-3,5),title="PT01 seizure 1",display=display)
}
