% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmap_frag}
\alias{heatmap_frag}
\title{Visualization functions (raw signal, fragility matrix)}
\usage{
heatmap_frag(
  frag,
  elecsoz,
  time_window = NULL,
  title = "Patient name seizure number",
  display = NULL
)
}
\arguments{
\item{frag}{Matrix or Fragility object. Either a matrix with row as Electrode names and Column as fragility index, or a Fragility object from \code{calc_adj_frag}}

\item{elecsoz}{Integer or string. Vector soz electrodes (for good electrodes)}

\item{time_window}{Numeric Vector of length 2. The time window to display at the x-axis}

\item{title}{String. Figure title}

\item{display}{Integer or string. Vector electrodes to display}
}
\value{
Heatmap plot of the fragility matrix with soz electrodes in blue in the bottom
}
\description{
plot fragility heatmaps with electrodes marked as soz colored
}
\examples{
data("pt01Epoch")
sozindex<-attr(pt01Epoch,"sozindex")
data("pt01Frag")
time_window <- c(-10,10)
display <- c(sozindex,77:80)
fragplot<-heatmap_frag(frag=pt01Frag,elecsoz=sozindex,time_window <- time_window,title="PT01 seizure 1",display=display)
fragplot

# use electrodes name for display and soz electrodes
data("pt01Epoch")
soznames<-attr(pt01Epoch,"soznames")
data("pt01Frag")
time_window <- c(-10,10)
display <- c(soznames,"MLT1","MLT2","MLT3","MLT4")
fragplot<-heatmap_frag(frag=pt01Frag,elecsoz=soznames,time_window <- time_window,title="PT01 seizure 1",display=display)
fragplot

# save plot to file with ggplot2
data("pt01Epoch")
data("pt01Frag")
sozindex<-attr(pt01Epoch,"sozindex")
time_window <- c(-10,10)
display <- c(sozindex,77:80)
pathplot <- "~"
title <- "PT01sz1"
resfile <- paste(pathplot,'/FragilityHeatMap',title,'.png',sep="")
fragplot<-heatmap_frag(frag=pt01Frag,elecsoz=sozindex,time_window=time_window,title=title,display=display)
fragplot
ggplot2::ggsave(resfile)

}
