% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_frag_distribution}
\alias{plot_frag_distribution}
\title{Plot Fragility time distribution for two electrodes group marked and non-marked as soz}
\usage{
plot_frag_distribution(statcurves, time_window_ictal, title = title)
}
\arguments{
\item{time_window_ictal}{time window of processed ictal iEEG}

\item{statcurves.}{List of four numeric vectors for mean and standard deviation statistics of the two groups}
}
\value{
plot fragility distribution
}
\description{
Plot Fragility time distribution for two electrodes group marked and non-marked as soz
}
\examples{
time_window_ictal=c(-3,5)
data("fragm3sp5s")
data("sozindex")
# compute fragility statistics evolution with time (mean and standard deviation) for soz and
# non soz groups
fragstat=frag_stat(frag=fragm3sp5s, elecsoz=sozindex)
fragstatcurve=fragstat[-1]
# plot the statistical results
pfragstat<-plot_frag_distribution(statcurves=fragstatcurve,time_window_ictal=time_window_ictal,title="PT01 seizure 1")
pfragstat
}
