% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_frag_quantile}
\alias{plot_frag_quantile}
\title{Plot Fragility time quantiles for two electrodes group marked as soz non marked as soz}
\usage{
plot_frag_quantile(qmatsozsozc, time_window_ictal, title = title)
}
\arguments{
\item{qmatsozsozc}{Numeric. quantile matrix for the two groups}

\item{time_window_ictal}{Numeric. time window of processed ictal iEEG}

\item{title}{String. Figure title}
}
\value{
Quantile plot
}
\description{
Plot Fragility time quantiles for two electrodes group marked as soz non marked as soz
}
\examples{
data("fragstat")
time_window_ictal=c(-3,5)
plot_frag_quantile( qmatsozsozc=fragstat[[1]], time_window_ictal=time_window_ictal,title=title)
}
