% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotFragDistribution}
\alias{plotFragDistribution}
\title{Plot Fragility time distribution for two electrodes group marked and non-marked as soz}
\usage{
plotFragDistribution(
  stat = NULL,
  timeWindow = NULL,
  title = "Average Fragility over time",
  cmeansoz = NULL,
  cmeansozc = NULL,
  csdsoz = NULL,
  csdsozc = NULL
)
}
\arguments{
\item{stat}{FragStat object, a FragStat object from \code{fragStat}, if specified, the arguments cmeansoz, cmeansozc, csdsoz, csdsozc will be ignored}

\item{timeWindow}{Numeric Vector of length 2. The time window to display at the x-axis}

\item{title}{String. Figure title}
}
\value{
plot fragility distribution
}
\description{
Plot Fragility time distribution for two electrodes group marked and non-marked as soz
}
\examples{
data("pt01Epoch")
sozindex<-attr(pt01Epoch,"sozindex")
# Load the precomputed fragility object
timeWindow <- c(-10,10)
data("pt01Frag")
# compute fragility statistics evolution with time (mean and standard deviation) for soz and
# non soz groups
pt01fragstat <- fragStat(frag=pt01Frag, elecSoz=sozindex)
# plot the statistical results
pfragstat<-plotFragDistribution(stat=pt01fragstat,timeWindow=timeWindow)
pfragstat
}
