% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotFragQuantile}
\alias{plotFragQuantile}
\title{Plot Fragility time quantiles for two electrodes group marked as soz non marked as soz}
\usage{
plotFragQuantile(
  FragStatObj,
  time_window = NULL,
  title = "Fragility Quantiles over time"
)
}
\arguments{
\item{FragStatObj}{Matrix or FragStat object, either a quantile matrix
for the two groups or a FragStat object from \code{frag_stat}}

\item{time_window}{Numeric Vector of length 2. The time window to display at the x-axis}

\item{title}{String. Figure title}
}
\value{
Quantile plot
}
\description{
Plot Fragility time quantiles for two electrodes group marked as soz non marked as soz
}
\examples{
time_window <- c(-10,10)
data("pt01Epoch")
sozindex<-attr(pt01Epoch,"sozindex")
data("pt01Frag")
# compute fragility statistics evolution with time (mean and standard deviation) for soz and
# non soz groups
pt01fragstat <- fragStat(frag=pt01Frag, elecSoz=sozindex)
plot_frag_quantile(FragStatObj=pt01fragstat, time_window=time_window)
}
