% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pt01Epoch}
\alias{pt01Epoch}
\alias{pt01Fragm1sp2s}
\title{Pt01 seizure 1 (-10:10s) around seizure onset}
\format{
pt01Epoch: A Matrix with 20001 rows (time points) and 84 columns (electrodes)

pt01Fragm1sp2s: A fragility object result of example 2 for
calc_adj_frag function help with and 84 columns (electrodes)
\link{-1:2}s around the seizure onset
}
\source{
Fragility Multi-Center Retrospective Study
(\href{https://openneuro.org/datasets/ds003029/versions/1.0.0}{OpenNeuro})
}
\usage{
## EEG data
data(pt01Epoch)

## pt01Fragm1sp2s Object
data(pt01Fragm1sp2s)
}
\description{
This data corresponds to the first seizure of patient PT01 from
the Fragility Data Set.
The data contains only the good channels.
It has been notch filtered and common average referenced in RAVE.
The full data (pt01Epoch) has been epoched -10:10s around the seizure onset.
The acquisition frequency is 1000 Hz
EcoG recording gathered in collaboration with the National Institue of Health
}
\keyword{datasets}
