% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigde.r
\name{ridge}
\alias{ridge}
\title{fit a generalized linear model to compute adjacency matrix A}
\usage{
ridge(xt, xtp1, lambda)
}
\arguments{
\item{xt}{matrix. iEEG time series for a given window,
with electrodes names as rows and time points as columns}

\item{xtp1}{matrix. the iEEG time serie at the next time point,
with electrodes names as rows and time points as columns}

\item{lambda}{Numeric Vector. A user supplied lambda sequence.}
}
\value{
adjacency matrix A
}
\description{
A x(t) = x(t+1)
}
