% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigde.r
\name{ridge}
\alias{ridge}
\title{fit a generalized linear model to compute adjacency matrix A}
\usage{
ridge(xt, xtp1, lambda, intercept = FALSE)
}
\arguments{
\item{xt}{matrix. iEEG time series for a given window}

\item{xtp1}{matrix. the iEEG time serie at the next time point}

\item{lambda}{Numeric. A user supplied lambda sequence.}

\item{intercept}{Boolean. Should intercept(s) be fitted (default=TRUE) or set to zero (FALSE)}

\item{iw}{}
}
\value{
adjacency matrix A
}
\description{
A x(t) = x(t+1)
}
