% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigde.r
\name{ridge}
\alias{ridge}
\title{fit a generalized linear model to compute adjacency matrix A}
\usage{
ridge(xt, xtp1, lambda, intercept = FALSE, iw)
}
\arguments{
\item{xt}{}

\item{xtp1}{}

\item{lambda}{}

\item{intercept}{}

\item{iw}{}
}
\value{
adjacency matrix A
}
\description{
A xt = xtp1
}
