% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigde.r
\name{ridgesearchlambdadichomotomy}
\alias{ridgesearchlambdadichomotomy}
\title{Ridge Regression for Electrode Readings}
\usage{
ridgesearchlambdadichomotomy(xt, xtp1, intercept = FALSE)
}
\arguments{
\item{xt}{matrix. iEEG time series for a given window,
with time points as rows and electrodes names as columns}

\item{xtp1}{matrix. the iEEG time serie at the next time point,
with time points as rows and electrodes names as columns}

\item{intercept}{Boolean. Should intercept(s) be fitted (default=TRUE) or set to zero (FALSE)}
}
\value{
adjacency matrix Afin with lambda as attribute
}
\description{
Ridge regression to compute matrix adjancency matrix A such as A xt = xtpt1
the lambda parmeter is found by dichotomy such that A is stable
(all eigenvalues have a norm less than one)
}
