% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Fragility.r
\name{[,Fragility-method}
\alias{[,Fragility-method}
\title{Subset a Fragility object}
\usage{
\S4method{[}{Fragility}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A Fragility object}

\item{i}{A logical vector or a numeric vector of indices to subset the electrodes}

\item{j}{A logical vector or a numeric vector of indices to subset the time windows}

\item{...}{Additional arguments (not used)}

\item{drop}{Additional arguments (not used)}
}
\value{
A new Fragility object with the subsetted data
}
\description{
Subset a Fragility object
}
