% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{valid_soz}
\alias{valid_soz}
\title{validate seizure onset data}
\usage{
valid_soz(ieegts, sozindex, soznames)
}
\arguments{
\item{ieegts}{Numeric. A matrix of iEEG time series x(t),
with time points as rows and electrodes names as columns}

\item{sozindex}{Integer. Vector soz electrodes}

\item{soznames}{Vector string. soz electrodes names}
}
\value{
boolean
}
\description{
validate seizure onset data
}
\examples{
\dontrun{
data("pt01Epochm1sp2s")
sozindex<-attr(pt01Epochm1sp2s,"sozindex")
soznames<-attr(pt01Epochm1sp2s,"soznames")
valid<-valid_soz(ieegts=pt01Epochm1sp2s,sozindex=sozindex,soznames=soznames)
}
}
