% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visuIEEGData}
\alias{visuIEEGData}
\title{Visualization of ictal iEEG}
\usage{
visuIEEGData(epoch)
}
\arguments{
\item{epoch}{Matrix or Epoch object. iEEG data matrix or Epoch object. If matrix, the row names are the electrode names and the column names are the time points}
}
\value{
A ggplot object
}
\description{
Visualization of ictal iEEG
}
\examples{
data("pt01EcoG")

## Visualize a subject of electrodes
sozIndex <- attr(pt01EcoG, "sozIndex")
display <- c(sozIndex, 77:80)

epoch <- Epoch(pt01EcoG)
visuIEEGData(epoch = epoch[display, ])
}
