% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visuiEEGdata}
\alias{visuiEEGdata}
\title{Visualization of ictal iEEG}
\usage{
visuiEEGdata(
  ieegts,
  time_window = NULL,
  title = "Patient name seizure number",
  display = NULL
)
}
\arguments{
\item{ieegts}{Matrix or Fragility object. Either a matrix of iEEG time
series x(t), with time points as rows and electrodes names as columns,
or a Fragility object from \code{calc_adj_frag}}

\item{time_window}{Numeric Vector of length 2. The time window to display at the x-axis}

\item{title}{String. Figure title}

\item{display}{Integer or string. Vector electrodes to display}
}
\value{
plot raw signal
}
\description{
Visualization of ictal iEEG
}
\examples{
data("pt01Epoch")
sozindex <- attr(pt01Epoch,"sozindex")

## We will use data from -1s to 2s around the seizure onset
pt01Epochm1sp2s<-pt01Epoch[9001:12000,]

display <- c(sozindex,77:80)
time_window <- c(-1,2)
iEEGplot<-visuiEEGdata(ieegts=pt01Epochm1sp2s,time_window=time_window,display=display)
iEEGplot
}
