% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visuiEEGdata}
\alias{visuiEEGdata}
\title{Visualization of ictal iEEG}
\usage{
visuiEEGdata(ieegts, scaling = 1e+06, display = NULL)
}
\arguments{
\item{ieegts}{Numeric. A matrix of iEEG time series x(t),
with time points as rows and electrodes names as columns}

\item{scaling}{Numeric. Scaling factor}

\item{display}{Integer or string. Vector electrodes to display}
}
\value{
plot raw signal
}
\description{
Visualization of ictal iEEG
}
\examples{
data("pt01Epochm3sp5s")
data("sozindex")
display=sozindex
visuiEEGdata(ieegts=PT01Epochm30sp30s,scaling=1000000, display=display)
}
