% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visuIEEGData}
\alias{visuIEEGData}
\title{Visualization of ictal iEEG}
\usage{
visuIEEGData(
  ieegts,
  timeRange = NULL,
  title = "Patient name seizure number",
  display = NULL
)
}
\arguments{
\item{ieegts}{Matrix or Fragility object. Either a matrix of iEEG time
series x(t), with time points as rows and electrodes names as columns,
or a Fragility object from \code{calc_adj_frag}}

\item{timeRange}{Numeric Vector of length 2. The time window to display at the x-axis}

\item{title}{String. Figure title}

\item{display}{Integer or string. Vector electrodes to display}
}
\value{
plot raw signal
}
\description{
Visualization of ictal iEEG
}
\examples{
data("pt01Epochm1sp2s")
sozIndex <- attr(pt01Epochm1sp2s,"sozIndex")
display <- c(sozIndex,77:80)
timeRange <- c(-1,2)
iEEGplot<-visuIEEGData(ieegts=pt01Epochm1sp2s,timeRange=timeRange,display=display)
iEEGplot
}
