% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visuiEEGdata}
\alias{visuiEEGdata}
\title{Visualization iEEGData}
\usage{
visuiEEGdata(ieegts, scaling, displayChannels)
}
\arguments{
\item{ieegts}{Numeric. A matrix of iEEG time series x(t),
with time points as rows and electrodes names as columns}

\item{scaling}{Scaling factor}

\item{displayChannels}{Channels to display}
}
\value{
plot raw signal
}
\description{
Visualization iEEGData
}
\examples{
data("PT01Epochm30sp30s")
data("ElectrodesDataPT01")
displayChannels=which(ElectrodesDataPT01$insoz==TRUE)
visuiEEGdata(ieegts=PT01Epochm30sp30s,1000000, displayChannels = displayChannels)
}
