% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visuiEEGdata}
\alias{visuiEEGdata}
\title{Visualization of ictal iEEG}
\usage{
visuiEEGdata(
  ieegts,
  time_window,
  title = "Patient name seizure number",
  display = NULL
)
}
\arguments{
\item{ieegts}{Numeric. A matrix of iEEG time series x(t),
with time points as rows and electrodes names as columns}

\item{time_window}{Numeric Vector. Fragility heatmap time window around seizure onset (s)}

\item{title}{String. Figure title}

\item{display}{Integer or string. Vector electrodes to display}
}
\value{
plot raw signal
}
\description{
Visualization of ictal iEEG
}
\examples{
data("pt01Epochm3sp5s")
sozindex<-attr(pt01Epochm3sp5s,"sozindex")
display=c(sozindex,77:80)
time_window=c(-3,5)
iEEGplot<-visuiEEGdata(ieegts=pt01Epochm3sp5s,time_window=time_window,display=display)
iEEGplot
}
